# Precompile some model examples reducing TTFX