--- a/src/Agents.jl
+++ b/src/Agents.jl
@@ -54,6 +54,8 @@ include("deprecations.jl")
 # visualizations (singleton methods for package extension)
 include("visualizations.jl")
 
+# precompile some example models to reduce TTFX#
+include("precompile.jl")
 
 # Update messages:
 using Scratch