This addresses a comment in https://github.com/JuliaDynamics/Agents.jl/issues/818 .

I think it's a very good improvement in TTFX with little side effects as I will illustrate:

Before pr:

`@time using Agents # -> 3.466778`

After pr

`@time using Agents # -> 3.798331`

so there is a little regression, however running this:

```julia
using Agents
models = [Models.flocking, Models.schelling, Models.zombies, Models.sir]

for m in models
    t = @time model, agent_step!, model_step! = m(); step!(model, agent_step!, model_step!, 10)
end
```

before:

```julia
  0.501026 seconds (1.27 M allocations: 87.205 MiB, 6.06% gc time, 99.95% compilation time)
  0.353901 seconds (1.15 M allocations: 76.176 MiB, 2.58% gc time, 99.96% compilation time)
  7.463354 seconds (13.40 M allocations: 879.895 MiB, 4.50% gc time, 99.23% compilation time: 4% of which was recompilation)
  1.325875 seconds (2.77 M allocations: 183.958 MiB, 3.19% gc time, 99.75% compilation time)
```

after:

```julia
  0.012826 seconds (2.43 k allocations: 129.844 KiB, 98.79% compilation time)
  0.011626 seconds (436 allocations: 33.000 KiB, 99.18% compilation time)
  0.200568 seconds (610.16 k allocations: 44.531 MiB, 10.97% gc time, 61.06% compilation time)
  0.007742 seconds (19.04 k allocations: 2.683 MiB, 61.97% compilation time)
```

So I think a win overall

Precompiling the models in the Models folder makes the examples models almost without latency which is good for users using the library for the first time