@Tortar 2023-07-13
mmmmhhhh I'm actually not sure that this is as good as it was intended to be since running https://julialang.github.io/PrecompileTools.jl/stable/#Seeing-what-got-precompiled we get e.g.

`MethodInstance for Agents.add_agent!(::Agents.UnremovableABM{Agents.ContinuousSpace{2, true, Float64, typeof(Agents.no_vel_update)}, Main.Agents.Models.Bird, Agents.Schedulers.Randomly, Nothing, Random.TaskLocalRNG}, ::Tuple{Float64, Float64}, ::Float64, ::Float64, ::Float64, ::Float64, ::Float64, ::Float64)`

which is actually still not reusable in other contexts, indeed for examples not precompiled (e.g. `predator_prey`) the time is very similar with or without precompilation.

The advantage of fast tutorial examples seems still something to consider, but don't know if worth the (even if little) regression in importing