@Datseris 2023-07-13
Yeah it's because practically all functions in Agents.jl dispatch on the Agent type, which is almost certainly different in each application. So pre-compiling is very unlikely to lead to any difference in user developed models...